requirements.txt
flask
scikit-learn
pandas
joblib
python-docx
pdfplumber
flask-session
pytest
PyPDF2